Change the resolution container to uint64

+++ entities/init_test.go
@@ -72,7 +72,7 @@ var (
 		RaceID:         1,
 		TwitterID:      "asdf",
 		HomePlanet:     "planet.GOP6720",
-		ScreenSize:     []uint16{1, 1},
+		ScreenSize:     []uint64{1, 1},
 		ScreenPosition: &vec2d.Vector{2, 2},
 	}
 	sun Sun = Sun{
